Refactored automaticallyplayed params

+++ GameProcessingService/CardEffectProcessors/Played/Wild/FairPlayEffectProcessor.cs
@@ -25,28 +25,35 @@ namespace GameProcessingService.CardEffectProcessors.Played.Wild
             var messagesToLog = new List<string>();
             var messageToLog = $"{moveParams.PlayerPlayed.User.Name} targeted {moveParams.PlayerTargeted.User.Name} with Fair Play. ";
 
-            var automaticallyTriggeredResultMagneticPolarity = _automaticallyTriggeredCardEffectProcessors.First(x => x.CardAffected == CardValue.MagneticPolarity).ProcessCardEffect(game, new AutomaticallyTriggeredParams(moveParams, messageToLog, null, 0));
+            var automaticallyTriggeredResultMagneticPolarity = _automaticallyTriggeredCardEffectProcessors.First(x => x.CardAffected == CardValue.MagneticPolarity).ProcessCardEffect(game, messageToLog, new AutomaticallyTriggeredParams() { MagneticPolarityParams = new AutomaticallyTriggeredMagneticPolarityParams(moveParams.TargetedCardColor, moveParams.PlayerPlayed, moveParams.PlayerTargeted) });
             messageToLog = automaticallyTriggeredResultMagneticPolarity.MessageToLog;
+            moveParams.PlayerTargeted = automaticallyTriggeredResultMagneticPolarity.MagneticPolaritySelectedPlayer;
 
-            var cardDifference = moveParams.PlayerTargeted.Cards.Count - moveParams.PlayerPlayed.Cards.Count;
+            var automaticallyTriggeredResultKeepMyHand = _automaticallyTriggeredCardEffectProcessors.First(x => x.CardAffected == CardValue.KeepMyHand).ProcessCardEffect(game, messageToLog, new AutomaticallyTriggeredParams() { KeepMyHandParams = new AutomaticallyTriggeredKeepMyHandParams(new List<Player>() { moveParams.PlayerTargeted }, moveParams.TargetedCardColor) });
+            messageToLog = automaticallyTriggeredResultKeepMyHand.MessageToLog;
 
-            if (cardDifference == 0)
+            if (automaticallyTriggeredResultKeepMyHand.PlayersWithoutKeepMyHand.Count == 1)
             {
-                messageToLog += "No effect. They had the same number of cards.";
+                var cardDifference = moveParams.PlayerTargeted.Cards.Count - moveParams.PlayerPlayed.Cards.Count;
+
+                if (cardDifference == 0)
+                {
+                    messageToLog += "No effect. They had the same number of cards.";
+                }
+                else if (cardDifference > 0)
+                {
+                    moveParams.PlayerTargeted.Cards.RemoveRange(0, cardDifference);
+                    messageToLog += $"{moveParams.PlayerTargeted.User.Name} discarded {cardDifference} cards. They had more cards.";
+                }
+                else
+                {
+                    var numberInPositiveValue = cardDifference * -1;
+                    _gameManager.DrawCard(game, moveParams.PlayerTargeted, numberInPositiveValue, false);
+                    messageToLog += $"{moveParams.PlayerTargeted.User.Name} must draw {numberInPositiveValue} cards. They had less cards.";
+                }
             }
-            else if (cardDifference > 0)
-            {
-                //targeted Player discards
-                moveParams.PlayerTargeted.Cards.RemoveRange(0, cardDifference);
-                messageToLog += $"{moveParams.PlayerTargeted.User.Name} discarded {cardDifference} cards. They had more cards.";
-            }
-            else
-            {
-                //targeted Player draws
-                var numberInPositiveValue = cardDifference * -1;
-                _gameManager.DrawCard(game, moveParams.PlayerTargeted, numberInPositiveValue, false);
-                messageToLog += $"{moveParams.PlayerTargeted.User.Name} must draw {numberInPositiveValue} cards. They had less cards.";
-            }
+
+
             messagesToLog.Add(messageToLog);
             return new MoveResult(messagesToLog);
         }
